✨ feat: 支持 goose 提供商并优化会话列表展示
- 新增 goose 提供商支持，更新文档和错误提示中的提供商列表
- 重构会话扫描逻辑，支持通过 `--include-empty` 参数显示空草稿会话
- 移除 `render_session_table` 中未使用的选中状态参数，简化表格渲染
- 调整 `--limit` 参数类型为 `u64` 并增加范围校验，确保值至少为 1
- 更新 `.codegraph/.gitignore` 规则，忽略目录内所有文件但保留 `.gitignore` 自身

+++ src/lib.rs
@@ -133,12 +133,12 @@ pub struct PsArgs {
 
 #[derive(Debug, Clone, Args)]
 pub struct SessionsArgs {
-    /// Filter by provider: claude, codex, cursor, gemini, opencode, pi, or omp
+    /// Filter by provider: claude, codex, cursor, gemini, goose, opencode, pi, or omp
     #[arg(long)]
     pub provider: Option<String>,
     /// Show only the most recently updated N sessions
-    #[arg(long)]
-    pub limit: Option<usize>,
+    #[arg(long, value_parser = clap::value_parser!(u64).range(1..))]
+    pub limit: Option<u64>,
     /// Emit stable machine-readable JSON
     #[arg(long)]
     pub json: bool,
